Add tests for BufferFactory. Also improve test code-reuse
